split post_transforms and review
