+++ sphinx_exercise/assets/html/exercise.css
