+++ sphinx_exercise/utils.py
@@ -1,10 +1,12 @@
 # Utility functions
+
 from sphinx.writers.latex import LaTeXTranslator
 from docutils import nodes as docutil_nodes
 
 
 def find_parent(env, node, parent_tag):
     """Find the nearest parent node with the given tagname."""
+
     while True:
         node = node.parent
         if node is None:
@@ -25,6 +27,7 @@ def find_parent(env, node, parent_tag):
 
 def get_node_number(self, node, typ) -> str:
     """Get the number for the directive node for HTML."""
+
     ids = node.attributes.get("ids", [])[0]
     if isinstance(self, LaTeXTranslator):
         docname = find_parent(self.builder.env, node, "section")
@@ -38,6 +41,7 @@ def get_node_number(self, node, typ) -> str:
 
 def has_math_child(node):
     """ Check if a parent node as a math child node. """
+
     for item in node:
         if isinstance(item, docutil_nodes.math):
             return True
@@ -46,6 +50,7 @@ def has_math_child(node):
 
 def get_refuri(node):
     """ Check both refuri and refid, to see which one is available. """
+
     id_ = ""
     if node.get("refuri", ""):
         id_ = node.get("refuri", "")
@@ -58,6 +63,7 @@ def get_refuri(node):
 
 def list_rindex(li, x) -> int:
     """Getting the last occurence of an item in a list."""
+
     for i in reversed(range(len(li))):
         if li[i] == x:
             return i
